implement individual photo pages
Create templates/photo.html
Create handler for individual images -- handleImagePage
Link to image page in templates/album.html

When the siteHandler can't find an album for the path, split the path
into to parts at the last "/" and see if the first part is an album. If
it is, then the second part is an image slug.

If the slug actually represents a valid object in this bucket, call
handleImagePage. Otherwise just redirect back to the album page.

We don't have to check paths split from other instances of the delimiter
because we know that the images in an album cannot contain the
delimiter.

Future work:

- I'm not sure how to check (or if we need to check) the existance of
  the image via the imgix CDN. Additionally, we should probably just
  cache whether this image exists like we do for album pages (or use
  that same cache)

- Add navigation buttons -- use the album contents to provide prev/next
  links to move through the album without returning to the album index.
  Potentially also include first/last links?

- Add JS to extract EXIF/IPTC metadata from the image and insert it into
  the page. This will let us display title/caption/etc. next to an
  image.  Need to look for and decide on a library to use, or roll our
  own if that's not too hard.